+++ android/src/main/kotlin/dev/brewkits/native_workmanager/NativeWorkmanagerPlugin+Graph.kt
@@ -6,8 +6,6 @@ import androidx.work.NetworkType
 import androidx.work.OneTimeWorkRequest
 import androidx.work.WorkContinuation
 import androidx.work.WorkManager
-import dev.brewkits.kmpworkmanager.background.data.KmpHeavyWorker
-import dev.brewkits.kmpworkmanager.background.data.KmpWorker
 import dev.brewkits.kmpworkmanager.background.data.NativeTaskScheduler
 import dev.brewkits.kmpworkmanager.background.domain.*
 import io.flutter.plugin.common.MethodCall